Allow source files without metadata
Sometimes we would like to fill the metadata of a resource from a PGT patch. For example, when the PGT is constructed with an "overlay"  template and the metadata of a resource should come from the input values of the overlay template.
In such cases the creation of the metadata in the source file with some placeholder values should not be necessary.

Moreover, if someone does not define metadata for the resource in the source file and then in the PGT they define metadata fields the current implementation panics, as the code accesses resourceMap["metadata"] without any check for the existence of the metadata field. A panic should not occur.

This is a proposal here to implement a solution that avoids the panic and that enables the creation of the metadata from the PGT.